Refactor recognition related code and add support to work with different models depending on the data source
