+++ app/core/recognition/downloader.ts
@@ -1,3 +1,4 @@
+import { SensingDataSource } from "~/core/mode";
 import CustomModelDownloadConditions = com.google.firebase.ml.modeldownloader.CustomModelDownloadConditions;
 import CustomModel = com.google.firebase.ml.modeldownloader.CustomModel;
 import FirebaseModelDownloader = com.google.firebase.ml.modeldownloader.FirebaseModelDownloader;
@@ -5,11 +6,16 @@ import DownloadType = com.google.firebase.ml.modeldownloader.DownloadType;
 import OnSuccessListener = com.google.android.gms.tasks.OnSuccessListener;
 import OnFailureListener = com.google.android.gms.tasks.OnFailureListener;
 
+const modelNameForDataSource = new Map([
+  [SensingDataSource.LOCAL_DEVICE, "smartphone-model"],
+  [SensingDataSource.PAIRED_DEVICE, "tug-test-full"]
+]);
+
 export class ModelDownloader {
 
   private modelFilePath: string;
 
-  constructor(private modelName: string) {
+  constructor(private dataSource: SensingDataSource) {
   }
 
   async getModelFilePath(): Promise<string> {
@@ -28,8 +34,9 @@ export class ModelDownloader {
   }
 
   private startDownloadTask(conditions: CustomModelDownloadConditions): Promise<CustomModel> {
+    const modelName = modelNameForDataSource.get(this.dataSource);
     const downloadTask = FirebaseModelDownloader.getInstance()
-      .getModel(this.modelName, DownloadType.LATEST_MODEL, conditions)
+      .getModel(modelName, DownloadType.LATEST_MODEL, conditions)
 
     return new Promise<CustomModel>((resolve, reject) => {
       downloadTask.addOnSuccessListener(
